@@ -26,17 +26,25 @@
   },
   "homepage": "https://github.com/Willian1661/calculate_project#readme",
   "devDependencies": {
+    "@babel/core": "^7.20.12",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "babel-loader": "^9.1.2",
     "bulma": "^0.9.4",
     "css-loader": "^6.7.3",
-    "eslint": "^8.30.0",
+    "eslint": "^8.33.0",
     "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-react": "^7.32.2",
     "html-webpack-plugin": "^5.5.0",
     "jest": "^29.3.1",
     "mini-css-extract-plugin": "^2.7.2",
     "node-sass": "^8.0.0",
     "prettier": "^2.8.1",
+    "react-devtools": "^4.27.1",
+    "sass": "^1.58.0",
     "sass-loader": "^13.2.0",
     "style-loader": "^3.3.1",
+    "url-loader": "^4.1.1",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.11.1"
@@ -44,6 +52,9 @@
   "dependencies": {
     "bulma-start": "^0.0.5",
     "eslint-plugin-jest": "^27.1.7",
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-test-renderer": "^18.2.0"
   }
 }
