@@ -1,6 +1,5 @@
 const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 
 module.exports = {
   mode: 'development',
@@ -21,26 +20,18 @@ module.exports = {
   module: {
     rules: [
       {
-      test: /\.scss$/,
-      use: [
-          MiniCssExtractPlugin.loader,
-          {
-            loader: 'css-loader'
-          },
-          {
-            loader: 'sass-loader',
-            options: {
-              sourceMap: true,
-            }
-          }
-        ]
+        test: /\.(js|jsx)$/,
+        exclude: /node_modules/,
+        use: {
+          loader: "babel-loader",
+        },
+      },
+    {
+      test: /\.(sa|sc|c)ss$/,
+      use: ["style-loader", "css-loader", "sass-loader"],
     },
       {
-        test: /\.css$/i,
-        use: ['style-loader', 'css-loader'],
-      },
-      {
-        test: /\.(png|svg|jpg|jpeg|gif)$/i,
+        test: /\.(png|svg|jpg|jpeg|gif|woff|woff2|eot|ttf)$/i,
         type: 'asset/resource',
       },
     ],
@@ -50,9 +41,6 @@ module.exports = {
     new HtmlWebpackPlugin({
       filename: 'index.html',
       template: './src/pages/index.html',
-    }),
-    new MiniCssExtractPlugin({
-      filename: 'mystyles.css'
     }),
   ],
 };
